requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
appdirs==1.4.4
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cssselect==1.2.0
dateparser==1.2.1
demjson3==3.0.6
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
facebook-scraper==0.2.59
facebook-sdk==3.1.0
fake-useragent==2.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
instaloader==4.14.1
lxml==5.3.1
multidict==6.1.0
outcome==1.3.0.post0
parse==1.20.2
propcache==0.2.1
pycparser==2.22
pyee==12.1.1
pyppeteer==0.0.25
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
w3lib==2.3.1
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
